pyproject.toml
[tool.poetry]
name = "aospa-ota"
version = "0.1.0"
description = "A backend for OTA"
authors = ["Amogh Lele <amolele@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
django-markdownfield = "^0.10.0"
uvicorn = "^0.17.6"
gunicorn = "^20.1.0"
whitenoise = "^6.0.0"
dj-database-url = "^0.5.0"
psycopg2 = "^2.9.3"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
